+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 - Fixed suggestions when a pattern contains upper case letters
 - Uniformed number delimeter insertion in all places, added a HTML escaping and fixed pattern matching
 - Fixed testing a string is URL in `struct` view to accept dashes in a domain name
-- Added `darkmode` attribute when dark mode is used on a preloader container
+- Added `darkmode` attribute on a preloader container when dark mode is used
+- Removed `dataField` parameter in `loadDataFromUrl()` method, this parameter might be passed via `options`
+- Added option `loadDataOptions` in preloader's config to pass it as `options` to load data methods
 
 ## 1.0.0-beta.66 (14-07-2022)